move utility fns to utils.py
